Update board.rs (#1)
* Update board.rs

you don't have to create an iterator over a string every time to get nth element. 

Good job by the way, and welcome to Fastly!

* Update board.rs

`any()` makes this much less verbose

* Update board.rs

Immediately dereferencing a reference is no-op and makes the code less clear.

* Update board.rs

oops, that flipped the logic. fixed

* Update board.rs

ok, this is embarrassing. lol. but this works